rename footer component
